[Refactor] Move command executor into Script layer
