+++ be/module_boundary_manifest.json
@@ -277,7 +277,6 @@
         "runtime/arrow_result_writer.h",
         "runtime/broker_mgr.h",
         "runtime/buffer_control_block.h",
-        "runtime/command_executor.h",
         "runtime/customized_result_writer.h",
         "runtime/data_stream_mgr.h",
         "runtime/data_stream_recvr.h",
@@ -1081,12 +1080,104 @@
       "remediation": "Keep Orchestration below Service; move transport-specific RPC handling to Service and lower reusable execution/runtime or ingestion primitives to their owning modules."
     },
     {
+      "id": "script",
+      "doc_label": "Script",
+      "summary": "Diagnostic script execution and command dispatch layer below Service, HttpService, Tools, and AgentServer, and above the remaining reusable BE modules.",
+      "owned_targets": ["Script"],
+      "owned_roots": ["be/src/script"],
+      "allowed_include_prefixes": [
+        "script/",
+        "base/",
+        "cache/",
+        "column/",
+        "common/",
+        "connector/",
+        "compute_env/",
+        "data_workflows/",
+        "exec/",
+        "exprs/",
+        "formats/",
+        "fs/",
+        "gen_cpp/",
+        "geo/",
+        "gutil/",
+        "io/",
+        "orchestration/",
+        "platform/",
+        "runtime/",
+        "serde/",
+        "storage/",
+        "types/",
+        "udf/",
+        "util/"
+      ],
+      "forbidden_include_prefixes": [
+        "service/",
+        "agent/",
+        "tools/",
+        "http/"
+      ],
+      "allowed_target_deps": [
+        "Base",
+        "Gutil",
+        "Common",
+        "Cache",
+        "IO",
+        "FSCore",
+        "FileSystem",
+        "Platform",
+        "Types",
+        "ColumnCore",
+        "ChunkCore",
+        "ColumnSortCore",
+        "RuntimeCore",
+        "RuntimeEnv",
+        "Runtime",
+        "FormatCore",
+        "FormatCsv",
+        "FormatJson",
+        "FormatAvro",
+        "FormatOrc",
+        "Formats",
+        "StoragePrimitive",
+        "StorageBase",
+        "Storage",
+        "ComputeEnv",
+        "DataWorkflows",
+        "ExprCore",
+        "Exprs",
+        "ExecPrimitive",
+        "ExecRuntime",
+        "ExecSchemaScannerCore",
+        "ExecSchemaScanners",
+        "ExecJoinCore",
+        "Exec",
+        "ConnectorPrimitive",
+        "Connector",
+        "ConnectorBenchmark",
+        "ConnectorElasticsearch",
+        "ConnectorMySQL",
+        "ConnectorBootstrap",
+        "ConnectorBuiltinRegistry",
+        "Orchestration",
+        "Serde",
+        "Udf",
+        "Util",
+        "Geo",
+        "StarRocksGen"
+      ],
+      "allowed_test_targets": ["script_test"],
+      "allowed_test_link_deps": ["Script"],
+      "remediation": "Keep Script below Service, HttpService, Tools, and AgentServer; lower reusable behavior can live in lower BE modules that Script is allowed to depend on."
+    },
+    {
       "id": "service",
       "doc_label": "Service",
-      "summary": "Shared service-layer target above runtime, cache, compute, and AgentServer without owning ServiceBE bootstrap code.",
+      "summary": "Shared service-layer target above Script, runtime, cache, compute, and AgentServer without owning ServiceBE bootstrap code.",
       "owned_targets": ["Service"],
       "allowed_include_prefixes": [
         "service/",
+        "script/",
         "agent/",
         "base/",
         "cache/",
@@ -1112,6 +1203,7 @@
       ],
       "forbidden_include_prefixes": ["service/service_be/", "service_be/"],
       "allowed_target_deps": [
+        "Script",
         "Runtime",
         "Orchestration",
         "RuntimeEnv",
