+++ be/src/script/command_executor.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "command_executor.h"
+#include "script/command_executor.h"
 
 #include <rapidjson/document.h>
