+++ be/src/service/internal_service.cpp
@@ -87,11 +87,11 @@
 #include "orchestration/runtime_filter_worker.h"
 #include "runtime/batch_write/batch_write_mgr.h"
 #include "runtime/closure_guard.h"
-#include "runtime/command_executor.h"
 #include "runtime/descriptors.h"
 #include "runtime/exec_env.h"
 #include "runtime/lookup_stream_mgr.h"
 #include "runtime/time_guard.h"
+#include "script/command_executor.h"
 #include "service/service_metrics.h"
 #include "storage/storage_engine.h"
 #include "storage/txn_manager.h"
