+++ be/test/script/command_executor_test.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "runtime/command_executor.h"
+#include "script/command_executor.h"
 
 #include <gtest/gtest.h>
